added dynamic PIR analog threshold detection reset
I had an issue where I had the analog PIR threshold set to 800 (minimum value to trigger camera), it was at that time idling around 140-160 analog. Now it's back down to 0-4 and peaks under 800. So the camera wasn't being triggered when the PIR detected motion and spiked. I was able to simply fix it by dropping the peak to 700, but decided to write a dynamic one capable of detecting and setting the max.