update
for as long as I've had this running since I made it, I always had to ssh after moving the files with filezilla since no root. Been meaning to add scp but haven't gotten around to it. This simple fix changes the ownership of the file, usually root:root now it will be pi:root so should be able to delete from filezilla means 1 login not 2